move sfn legacy provider into legacy package
